XIVY-2942 refactor: move abstract page objects
