+++ integrations/standalone/tests/pageobjects/abstract/Switch.ts
